works with one prayer method - enough for MVP

+++ app/src/main/java/com/alphaomardiallo/duwagol/common/domain/model/prayerTimesMethods/ParamsDS.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ParamsDS(
-    val Fajr: Int,
+    val Fajr: Double,
     val Isha: String,
 )
